% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gen_code}
\alias{gen_code}
\title{Translation dictionary amino acid <-> nucleotide codon}
\format{A data frame with 64 rows and  variables:
\describe{
  \item{aa}{amino acid}
  \item{codon}{nucleotide codon}
}}
\source{
\url{https://www.genscript.com/tools/codon-table}
}
\usage{
gen_code
}
\description{
A dataframe containing a mapping from each of 64 codons
to amino acids.
}
\keyword{datasets}
