% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hotspot_df}
\alias{hotspot_df}
\title{Hotspot dataframe for SHM}
\format{A data frame with 1024 rows and  variables:
\describe{
  \item{pattern}{amino acid}
  \item{toA}{probability of mutation to adenine}
  \item{toC}{probability of mutation to cytosine}
  \item{toG}{probability of mutation to guanine}
  \item{toT}{probability of mutation to thymine}
  \item{Source}{source of probability}
}}
\source{
\url{https://cran.r-project.org/package=AbSim}
}
\usage{
hotspot_df
}
\description{
A dataframe containing mutation probabilities for every possible 5mer pattern
}
\keyword{datasets}
