% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_seqs_exclusion.R
\name{hub_seqs_exclusion}
\alias{hub_seqs_exclusion}
\title{Deletes top hub sequences from repertoire, changing the network architecture.}
\usage{
hub_seqs_exclusion(repertoire, top_x = 0.005, report = FALSE)
}
\arguments{
\item{repertoire}{An annotated AIRR compliant repertoire.

(http://docs.airr-community.org/en/latest/)}

\item{top_x}{Determines what percentage of hub sequences get excluded

(Default: 0.005, i.e. Top 0.5 percent)}

\item{report}{If TRUE saves a csv file containing the excluded sequences.}
}
\value{
Repertoire reduced by hub sequence (new network architecture)
}
\description{
Deletes top hub sequences from repertoire, changing the network architecture.
}
\examples{
\dontrun{
repertoire <- immuneSIM(number_of_seqs = 50,species = "mm",receptor = "ig", chain = "h")
hub_seqs_exclusion(repertoire, top_x = 0.005,report = FALSE)
}
}
