% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insertions_and_deletion_lengths_df}
\alias{insertions_and_deletion_lengths_df}
\title{Dataframe containing insertion sequences and deletion lengths}
\format{A data frame with 500000 rows and variables:
\describe{
  \item{n1}{np1 insertions}
  \item{n2}{np2 insertions}
  \item{del_v}{lengths of V gene deletions}
  \item{del_d_5}{lengths of 5' end D gene deletions}
  \item{del_d_3}{lengths of 3' end D gene deletions}
  \item{del_j}{lengths of J gene deletions}
}}
\source{
\url{https://doi.org/10.1016/j.celrep.2017.04.054}
}
\usage{
insertions_and_deletion_lengths_df
}
\description{
A dataframe containing all insertions and deletions
observed in experimental data (pooled across all samples, Greiff, 2017)
This dataframe is a subset of the dataframe used in the application note.
The original dataframe which contains 11363603 rows can be downloaded
from:
}
\details{
https://github.com/GreiffLab/immuneSIM or using the provided
function: load_insdel_data()
}
\keyword{datasets}
