% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{length_dist_simulation}
\alias{length_dist_simulation}
\title{Vector containing VDJ length distributions}
\format{A vector with 10000 entries:
\describe{
  \item{length}{VDJ nucleotide lengths sampled from murine naive
  follicular B-cell data, Greiff 2017}
}}
\source{
\url{https://doi.org/10.1016/j.celrep.2017.04.054}
}
\usage{
length_dist_simulation
}
\description{
A vector containing 10000 VDJ lengths for simulating
of fully random sequences (independent of germline genes)
}
\keyword{datasets}
