% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_example_repertoires}
\alias{list_example_repertoires}
\title{Example repertoires}
\format{A list with 2 entries:
\describe{
  \item{example_repertoire_A}{Repertoire simulated using standard parameters (A)}
  \item{example_repertoire_A}{Repertoire simulated using standard parameters (B)}
}}
\source{
{
   \url{https://immunesim.readthedocs.io}
}
}
\usage{
list_example_repertoires
}
\description{
A list containing two example repertoires (100 sequences each)
simulated with immuneSIM using default parameters.
These repertoires are used in the examples.
}
\keyword{datasets}
