% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_insdel_data.R
\name{load_insdel_data}
\alias{load_insdel_data}
\title{Loads full insertion/deletion data from GitHub}
\usage{
load_insdel_data()
}
\value{
Dataframe containing insertions and deletions (11363603 rows, 6 columns)
}
\description{
Loads full insertion/deletion data from GitHub
}
\examples{
\donttest{
full_insertions_and_deletion_df <- load_insdel_data()
}
}
