% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{one_spot_df}
\alias{one_spot_df}
\title{One Spot}
\format{A dataframe with 32 entries:
\describe{
  \item{pattern}{amino acid}
  \item{toA}{probability of mutation to adenine}
  \item{toC}{probability of mutation to cytosine}
  \item{toG}{probability of mutation to guanine}
  \item{toT}{probability of mutation to thymine}
  \item{Source}{source of probability}
}}
\source{
{
   \url{https://cran.r-project.org/package=AbSim}

   \url{https://doi.org/10.1093/bioinformatics/btx533}
}
}
\usage{
one_spot_df
}
\description{
A dataframe containing a mutation probabilities to base per 5mer
(inherited from AbSim package)
}
\keyword{datasets}
