% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_report_repertoire.R
\name{plot_report_repertoire}
\alias{plot_report_repertoire}
\title{Plots main repertoire features (length distribution,amino acid frequencies and VDJ usage)}
\usage{
plot_report_repertoire(repertoire, output_folder = "")
}
\arguments{
\item{repertoire}{An annotated AIRR-compliant immuneSIM repertoire.

(http://docs.airr-community.org/en/latest/)}

\item{output_folder}{String containing full path of desired output folder. If empty figures will be output to current path.}
}
\value{
TRUE (plots saved as pdfs into subfolder 'figures')
}
\description{
Plots main repertoire features (length distribution,amino acid frequencies and VDJ usage)
}
\examples{
\dontrun{
repertoire <- immuneSIM(number_of_seqs = 10,species = "mm",receptor = "ig", chain = "h")
plot_report_repertoire(repertoire,output_folder="")
}
}
