% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_report_repertoire.R
\name{plot_report_repertoire}
\alias{plot_report_repertoire}
\title{Plots main repertoire features (length distribution,amino acid frequencies and VDJ usage)}
\usage{
plot_report_repertoire(repertoire, output_dir = "", verbose = TRUE)
}
\arguments{
\item{repertoire}{An annotated AIRR-compliant immuneSIM repertoire.

(http://docs.airr-community.org/en/latest/)}

\item{output_dir}{String containing full path of desired output folder. If empty figures will be output in tempdir().}

\item{verbose}{Determines whether messages on plot locations are output to user. (Default: TRUE)}
}
\value{
TRUE (plots saved as pdfs into subfolder 'figures')
}
\description{
Plots main repertoire features (length distribution,amino acid frequencies and VDJ usage)
}
\examples{
repertoire <- list_example_repertoires[["example_repertoire_A"]]
plot_report_repertoire(repertoire,output_dir="")
}
